# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface StudioWorkspaceState { [key: string]: unknown }

export interface StudioWorkspaceInput {
  state: StudioWorkspaceState;
  /**
     * @minimum 1
     * @nullable
     */
  expectedVersion: number | null;
}

export interface StudioWorkspaceResponse {
  state: StudioWorkspaceState;
  /** @minimum 1 */
  version: number;
}

export interface StudioWorkspaceConflict {
  error: string;
  current: StudioWorkspaceResponse;
}

export interface StudioLegalProfileInput {
  /** @maxLength 200 */
  registeredName: string;
  /** @maxLength 200 */
  tradingName: string;
  /** @maxLength 120 */
  country: string;
  /** @maxLength 1000 */
  registeredAddress: string;
  /** @maxLength 320 */
  privacyEmail: string;
  /** @maxLength 500 */
  website: string;
}

export interface StudioLegalProfileResponse {
  registeredName: string;
  tradingName: string;
  country: string;
  registeredAddress: string;
  privacyEmail: string;
  website: string;
  /** @nullable */
  updatedAt: string | null;
  isComplete: boolean;
  canEdit: boolean;
}

export type EmailProvider = typeof EmailProvider[keyof typeof EmailProvider];


export const EmailProvider = {
  google: 'google',
  microsoft: 'microsoft',
} as const;

export interface EmailProviderInfo {
  provider: EmailProvider;
  configured: boolean;
  redirectUri: string;
}

export interface EmailConnection {
  provider: EmailProvider;
  email: string;
  connectedAt: string;
}

export interface EmailConnectionsResponse {
  providers: EmailProviderInfo[];
  connections: EmailConnection[];
}

export interface EmailAuthorizationResponse {
  authorizationUrl: string;
}

export type EmailAnalysisSignal = typeof EmailAnalysisSignal[keyof typeof EmailAnalysisSignal];


export const EmailAnalysisSignal = {
  enquiry: 'enquiry',
  budget: 'budget',
  project: 'project',
  meeting: 'meeting',
  invoice: 'invoice',
  follow_up: 'follow_up',
} as const;

export interface EmailAnalysisCandidate {
  email: string;
  name: string;
  /** @minimum 1 */
  messageCount: number;
  lastMessageAt: string;
  latestSubject: string;
  signals: EmailAnalysisSignal[];
}

export interface EmailAnalysisBatchInput {
  /** @nullable */
  cursor?: string | null;
}

export interface EmailAnalysisBatchResponse {
  candidates: EmailAnalysisCandidate[];
  /** @nullable */
  cursor: string | null;
  /** @minimum 0 */
  scannedCount: number;
  /**
     * @minimum 0
     * @nullable
     */
  totalCount: number | null;
  complete: boolean;
}

export interface ApiError {
  error: string;
}

export type CompleteGoogleEmailAuthorizationParams = {
code?: string;
state?: string;
error?: string;
};

export type CompleteMicrosoftEmailAuthorizationParams = {
code?: string;
state?: string;
error?: string;
};