# lib/api-zod/src/generated/types/emailAnalysisBatchResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmailAnalysisCandidate } from './emailAnalysisCandidate';

export interface EmailAnalysisBatchResponse {
  candidates: EmailAnalysisCandidate[];
  /** @nullable */
  cursor: string | null;
  /** @minimum 0 */
  scannedCount: number;
  /**
     * @minimum 0
     * @nullable
     */
  totalCount: number | null;
  complete: boolean;
}