# lib/api-zod/src/generated/types/emailAnalysisCandidate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmailAnalysisSignal } from './emailAnalysisSignal';

export interface EmailAnalysisCandidate {
  email: string;
  name: string;
  /** @minimum 1 */
  messageCount: number;
  lastMessageAt: Date;
  latestSubject: string;
  signals: EmailAnalysisSignal[];
}