# lib/api-zod/src/generated/types/emailAnalysisSignal.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type EmailAnalysisSignal = typeof EmailAnalysisSignal[keyof typeof EmailAnalysisSignal];


export const EmailAnalysisSignal = {
  enquiry: 'enquiry',
  budget: 'budget',
  project: 'project',
  meeting: 'meeting',
  invoice: 'invoice',
  follow_up: 'follow_up',
} as const;