# lib/api-zod/src/generated/types/emailConnection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmailProvider } from './emailProvider';

export interface EmailConnection {
  provider: EmailProvider;
  email: string;
  connectedAt: Date;
}