# lib/api-zod/src/generated/types/emailConnectionsResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmailConnection } from './emailConnection';
import type { EmailProviderInfo } from './emailProviderInfo';

export interface EmailConnectionsResponse {
  providers: EmailProviderInfo[];
  connections: EmailConnection[];
}