# lib/api-zod/src/generated/types/emailProviderInfo.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { EmailProvider } from './emailProvider';

export interface EmailProviderInfo {
  provider: EmailProvider;
  configured: boolean;
  redirectUri: string;
}